requirements.txt
Faker==15.0.0
google-cloud==0.34.0
Jinja2==3.1.2
google-cloud-pubsub
ratelimit
